+++ src/TgBotLib/Bot.php
@@ -2690,4 +2690,26 @@
 
             return true;
         }
+
+        /**
+         * Use this method to change search keywords assigned to a regular or custom emoji sticker.
+         * The sticker must belong to a sticker set created by the bot. Returns True on success.
+         *
+         * @param string $sticker File identifier of the sticker
+         * @param array $keywords A JSON-serialized list of 0-20 search keywords for the sticker with total length of up to 64 characters
+         * @param array $options Optional parameters
+         * @return bool
+         * @throws TelegramException
+         * @link https://core.telegram.org/bots/api#setstickerkeywords
+         * @noinspection PhpUnused
+         */
+        public function setStickerKeywords(string $sticker, array $keywords, array $options=[]): bool
+        {
+            $this->sendRequest('setStickerKeywords', array_merge([
+                'sticker' => $sticker,
+                'keywords' => $keywords
+            ], $options));
+
+            return true;
+        }
     }